+++ src/domain/core/src/services/sync_service.rs
@@ -10,7 +10,6 @@
 use std::sync::Arc;
 
 use internal_error::{BoxedError, InternalError};
-use odf::dataset::InvalidIntervalError;
 use thiserror::Error;
 use url::Url;
 
@@ -249,7 +248,9 @@ pub enum SyncError {
     #[error(transparent)]
     DestinationAhead(#[from] DestinationAheadError),
     #[error(transparent)]
-    InvalidInterval(#[from] InvalidIntervalError),
+    InvalidInterval(#[from] odf::dataset::InvalidIntervalError),
+    #[error(transparent)]
+    OverwriteSeedBlock(#[from] OverwriteSeedBlockError),
     #[error(transparent)]
     Corrupted(#[from] CorruptedSourceError),
     #[error("Dataset was updated concurrently")]
@@ -418,5 +419,11 @@ impl From<IpfsAddError> for SyncError {
 pub struct UnsupportedIpfsStorageTypeError {
     pub url: Url,
 }
+
+////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+#[derive(Error, Clone, PartialEq, Eq, Debug)]
+#[error("Overwriting dataset id or type is restricted")]
+pub struct OverwriteSeedBlockError {}
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////