+++ src/domain/datasets/domain/src/use_cases/create_dataset_use_case.rs
@@ -82,6 +82,13 @@ pub enum CreateDatasetError {
     EmptyDataset,
 
     #[error(transparent)]
+    Access(
+        #[from]
+        #[backtrace]
+        odf::AccessError,
+    ),
+
+    #[error(transparent)]
     NameCollision(#[from] NameCollisionError),
 
     #[error(transparent)]